requirements.txt
django==2.2.7
Pillow==2.2.1
django-crispy-forms==1.0.0
django-jet==1.0.8
django-phone-field==1.8.0
djangorestframework==3.10.3
django-phonenumber-field==3.0.1
phonenumbers==8.10.23